sql/app/requirements.txt
flask
flask-login
pandas
psycopg2-binary
werkzeug
joblib
scikit-learn
requests
Flask-Limiter
pyzabbix
python-dotenv
gunicorn